EmailTemplateFix

* fixed template constraint on delete

* delete fix for EMail Template

* removed exception while getting email structure

* throwing error message for delete action in email Template


Approved-by: Logesh Kannan K
Approved-by: Shanmuga Boopathy k

+++ src/main/java/com/facilio/emailtemplate/command/DeleteEmailStructureCommand.java
@@ -4,7 +4,12 @@ import com.facilio.bmsconsole.templates.Template;
 import com.facilio.bmsconsole.util.TemplateAPI;
 import com.facilio.command.FacilioCommand;
 import com.facilio.constants.FacilioConstants;
+import com.facilio.db.builder.GenericSelectRecordBuilder;
+import com.facilio.db.criteria.CriteriaAPI;
+import com.facilio.db.criteria.operators.NumberOperators;
 import com.facilio.emailtemplate.context.EMailStructure;
+import com.facilio.modules.FieldFactory;
+import com.facilio.modules.ModuleFactory;
 import org.apache.commons.chain.Context;
 
 public class DeleteEmailStructureCommand extends FacilioCommand {
@@ -20,8 +25,25 @@ public class DeleteEmailStructureCommand extends FacilioCommand {
         }
 
         EMailStructure emailStructure = (EMailStructure) template;
+        if (getEmailTemplate(emailStructure.getId())){
+            throw new IllegalArgumentException("Email Template in use");
+        }
         TemplateAPI.deleteTemplate(emailStructure.getId());
 
         return false;
+    }
+
+    private boolean getEmailTemplate(long id) throws Exception {
+
+        GenericSelectRecordBuilder selectRecordBuilder = new GenericSelectRecordBuilder()
+                .table(ModuleFactory.getEMailTemplatesModule().getTableName())
+                .select(FieldFactory.getEMailTemplateFields())
+                .andCondition(CriteriaAPI.getCondition("EMAIL_STRUCTURE_ID","emailStructureId", String.valueOf(id), NumberOperators.EQUALS));
+
+        if(selectRecordBuilder.get().size() > 0){
+            return true;
+        }
+        return false;
+
     }
 }
